Clean-up code and reformat files
