+++ utils/plot.py
@@ -1,0 +1,45 @@
+import matplotlib
+import matplotlib.pyplot as plt
+import numpy as np
+
+matplotlib.use('TkAgg')
+
+results ={"train_loss": [], "test_loss": [], "test_accuracy": [7.6761057804820965, 9.150479756611281, 10.016381933068102, 13.526796161947109, 16.452141352679615, 18.41797332085186, 20.898666042593025, 22.49005382635151, 23.941025040954834, 25.111163117247834, 25.626023870816756, 25.391996255558155, 26.866370231687338, 27.825883454247602, 28.153522115609643, 31.312894921600748, 29.885326468523285, 29.534285045635386, 30.37678446056635, 35.68921132693658, 41.58670723145331, 43.178095015211795, 44.30142756845308, 41.96115141586707, 39.85490287853967, 40.65059677041891, 44.48864966065996, 45.00351041422888, 48.95857711209923, 48.794757781418205, 51.883922302831735, 54.50503159372806, 56.56447460800374, 59.98127779077931, 60.7535689211327, 59.23238942195179, 61.081207582494734, 61.362040720805055, 64.8490521881582, 65.08307980341681, 68.17224432483033, 70.93377018488182, 72.43154692253685, 73.29744909899368, 73.74210156798502, 75.4505031593728, 75.87175286683829, 76.33980809735549, 76.9716826585537, 77.22911303533817, 77.41633512754505, 76.87807161245027, 77.36952960449334, 78.30564006552773, 79.52258366487246, 80.9033465948982, 81.605429440674, 81.67563772525158, 81.55862391762228, 81.20758249473438, 81.16077697168267, 79.96723613386379, 79.92043061081208, 81.06716592557922, 81.93306810203605, 82.14369295576878, 83.17341446290662, 83.40744207816522, 83.5946641703721, 84.15633044699274, 84.08612216241517, 83.92230283173414, 84.5307746314065, 84.83501053124269, 86.09875965363914, 85.58389890007021, 86.84764802246666, 87.33910601450971, 87.36250877603557, 87.01146735314767, 86.89445354551837, 86.68382869178564, 87.26889772993214, 88.01778609875966, 88.41563304469928, 89.632576644044, 90.14743739761292, 90.38146501287152, 90.87292300491458, 90.54528434355254, 90.63889538965599, 90.9197285279663, 91.01333957406975, 91.1537561432249, 91.48139480458694, 91.59840861221625, 91.76222794289727, 91.87924175052656, 91.59840861221625, 91.5047975661128, 90.96653405101802, 91.45799204306108, 91.55160308916452, 91.69201965831968, 92.04306108120758, 91.6218113737421, 91.5047975661128, 92.01965831968172, 92.11326936578516, 92.2068804118886, 92.41750526562134, 92.51111631172478, 92.48771355019892, 92.3238942195179, 92.41750526562134, 92.76854668850925, 93.04937982681956, 93.09618534987129, 93.18979639597472, 93.18979639597472, 93.35361572665575, 93.21319915750058, 93.6110461034402, 93.65785162649193, 93.75146267259537, 93.68125438801779, 93.84507371869881, 93.89187924175053, 93.9620875263281, 93.93868476480225, 93.91528200327639, 93.89187924175053, 94.10250409548327, 94.03229581090568, 93.98549028785396, 94.35993447226772, 94.38333723379358, 94.38333723379358, 94.45354551837117, 94.43014275684531, 94.38333723379358, 94.1961151415867, 94.59396208752632, 94.5471565644746, 94.6173648490522, 94.57055932600046, 94.71097589515563, 94.64076761057805, 94.66417037210391, 84.60098291598409, 89.25813245963023, 91.55160308916452, 93.77486543412122, 94.31312894921601, 94.47694827989703, 94.31312894921601, 94.17271238006084, 94.47694827989703, 93.93868476480225, 87.01146735314767, 90.3346594898198, 93.28340744207817, 94.38333723379358, 94.35993447226772, 94.64076761057805, 94.66417037210391, 94.80458694125907, 94.64076761057805, 94.28972618769015, 87.76035572197519, 91.59840861221625, 93.21319915750058, 94.47694827989703, 94.43014275684531, 94.7811841797332, 94.82798970278493, 94.68757313362977, 94.71097589515563, 94.66417037210391, 88.39223028317342, 92.60472735782822, 94.26632342616429, 95.10882284109525, 95.48326702550901, 95.45986426398315, 94.94500351041422, 94.80458694125907, 94.5471565644746, 94.57055932600046, 88.86028551369061, 92.01965831968172, 94.89819798736251, 95.83430844839691, 96.20875263281067, 96.3725719634917, 95.74069740229348, 95.53007254856072, 95.15562836414696, 95.1322256026211, 89.65597940556985], "train_accuracy": [7.698607698607699, 11.216429699842022, 14.440734557595993, 12.5, 15.17077045274027, 19.38943894389439, 21.977186311787072, 27.052489905787347, 28.441127694859038, 25.680580762250454, 25.554705432287683, 26.005888125613346, 23.613312202852615, 33.333333333333336, 24.591738712776177, 30.529595015576323, 41.66001596169194, 35.81800580832527, 28.91640866873065, 29.025191675794087, 38.28289936664321, 46.59468438538206, 39.10761154855643, 48.93617021276596, 42.91539245667686, 34.65458663646659, 45.37107880642693, 46.04810996563574, 52.46636771300449, 55.53677092138631, 50.90047393364929, 54.716981132075475, 61.980830670926515, 57.483246463142216, 62.04778156996587, 69.0677966101695, 64.10958904109589, 66.9250645994832, 57.72690106295993, 69.14660831509846, 71.14661654135338, 65.6437125748503, 73.92739273927393, 81.09375, 76.036866359447, 76.94753577106518, 82.13592233009709, 76.9543973941368, 75.4880694143167, 78.35497835497836, 80.0, 80.45302013422818, 77.43668457405987, 76.44386761842959, 74.05119152691968, 80.6390977443609, 81.49498632634457, 84.49920508744039, 80.25862068965517, 84.0499306518724, 81.8958818958819, 87.1405750798722, 83.34630350194553, 78.38461538461539, 83.41264445955133, 84.2433697347894, 80.71495766698024, 85.36139066788655, 84.10364145658264, 82.74231678486997, 85.96078431372548, 86.82634730538922, 84.01135288552507, 86.1851332398317, 86.63446054750402, 88.44290657439447, 87.30325288562435, 89.71807628524047, 84.54545454545455, 87.93103448275862, 88.75181422351234, 88.74656907593778, 88.80341880341881, 88.94658753709199, 87.64510779436152, 86.04878048780488, 92.04152249134948, 90.9686817188638, 93.17038102084831, 92.50398724082935, 91.58600148920327, 90.61801935964259, 93.26424870466322, 92.36526946107784, 91.79039301310044, 94.0, 93.6046511627907, 92.25806451612904, 94.02887139107611, 92.08234362628662, 92.43119266055047, 92.53859348198971, 92.14002642007927, 91.1685994647636, 92.9266136162688, 94.36065573770492, 92.4597364568082, 91.1512027491409, 92.97173414820473, 93.86973180076629, 92.55765199161425, 93.18801089918256, 93.72909698996655, 94.95967741935483, 94.29429429429429, 92.43498817966903, 94.12878787878788, 93.83791024782317, 94.62044231918709, 94.27934621099554, 94.53499520613614, 94.60981496379726, 93.67859862909368, 94.88188976377953, 94.99509322865555, 95.50870760769936, 94.49866903283052, 96.15384615384616, 95.52656104380242, 94.41201000834029, 94.52523524379812, 95.33508541392904, 95.39094650205762, 94.02756508422665, 94.2051683633516, 95.05915100904663, 96.22113365990202, 95.80246913580247, 93.6741767764298, 95.3168044077135, 94.22572178477691, 94.7098976109215, 95.43178973717147, 95.35055350553506, 95.77329490874159, 94.5359930615785, 94.40267335004177, 96.01203912716328, 95.7201646090535, 85.03162333099087, 88.30935251798562, 93.75, 94.09594095940959, 93.66144668158091, 95.20504731861199, 94.79025710419486, 94.68085106382979, 94.7002606429192, 93.5930735930736, 87.48019017432647, 93.60426382411725, 92.8395061728395, 95.49815498154982, 96.64429530201342, 94.51718494271685, 93.54581673306772, 95.25521207764199, 94.83734087694484, 93.82064807837227, 91.8018018018018, 91.83370452858203, 95.91836734693878, 94.98970487302677, 94.43521594684385, 94.78957915831663, 95.31396438612933, 96.3917525773196, 94.1715237302248, 94.4649446494465, 87.99424874191229, 91.41891891891892, 94.80614484272128, 94.43155452436194, 94.98257839721255, 95.33187454412837, 95.20078354554359, 94.42993907745866, 94.84797297297297, 95.34883720930233, 89.88679245283019, 93.31941544885177, 93.75987361769353, 95.19650655021834, 96.17021276595744, 96.3773069036227, 96.12468407750632, 97.18111346018323, 96.21668099742047, 95.16757599376461, 91.25], "backdoor_test_loss": [], "backdoor_test_accuracy": [1.1233325532412826, 2.691317575473906, 4.189094313128949, 10.367423355956003, 17.55207114439504, 19.19026445120524, 19.939152820032763, 24.549496840627192, 19.37748654341212, 19.166861689679383, 20.290194242920666, 24.783524455885793, 24.198455417739293, 18.18394570559326, 11.958811139714486, 12.450269131757548, 10.578048209688744, 7.020828457758015, 6.622981511818395, 13.409782354317809, 18.90943131289492, 12.263047039550667, 14.345892815352212, 9.6653405101802, 5.8038848584132925, 8.518605195413059, 8.89304937982682, 7.371869880645916, 10.718464778843904, 7.371869880645916, 8.89304937982682, 11.584366955300725, 11.514158670723145, 12.263047039550667, 12.122630470395507, 11.397144863093844, 12.005616662766206, 12.309852562602387, 13.85443482330915, 13.409782354317809, 15.141586707231454, 16.030891645214137, 16.709571729464077, 18.02012637491224, 17.71589047507606, 17.99672361338638, 18.254153990170842, 18.11373742101568, 17.99672361338638, 17.99672361338638, 17.76269599812778, 17.435057336765738, 17.154224198455417, 17.247835244558857, 17.528668382869178, 17.83290428270536, 17.71589047507606, 17.83290428270536, 17.388251813714017, 17.435057336765738, 17.6924877135502, 17.528668382869178, 17.528668382869178, 17.50526562134332, 17.73929323660192, 17.83290428270536, 17.8095015211795, 17.97332085186052, 17.97332085186052, 17.97332085186052, 17.87970980575708, 17.83290428270536, 17.83290428270536, 17.8095015211795, 17.90311256728294, 17.99672361338638, 18.02012637491224, 18.11373742101568, 18.11373742101568, 18.09033465948982, 18.11373742101568, 18.11373742101568, 18.11373742101568, 18.11373742101568, 18.11373742101568, 18.11373742101568, 18.11373742101568, 18.11373742101568, 18.13714018254154, 18.13714018254154, 18.11373742101568, 18.11373742101568, 18.1605429440674, 18.13714018254154, 18.11373742101568, 18.11373742101568, 18.13714018254154, 18.09033465948982, 18.09033465948982, 18.09033465948982, 18.13714018254154, 18.20734846711912, 18.20734846711912, 18.11373742101568, 18.11373742101568, 18.18394570559326, 18.20734846711912, 18.20734846711912, 18.23075122864498, 18.20734846711912, 18.20734846711912, 18.23075122864498, 18.23075122864498, 18.20734846711912, 18.20734846711912, 18.20734846711912, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.1605429440674, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.1605429440674, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.18394570559326, 18.20734846711912, 43.22490053826351, 42.710039784694594, 43.388719868944534, 43.43552539199626, 41.09524923941025, 39.691083547858646, 37.98268195647086, 36.74233559560028, 35.057336765738356, 35.150947811841796, 60.80037444418441, 59.06857009127077, 57.14954364615024, 54.879475778141824, 54.41142054762462, 51.86051954130588, 49.26281301193541, 48.27989702784929, 46.38427334425462, 45.026913175754736, 60.82377720571027, 59.60683360636555, 58.76433419143459, 57.032529838520944, 56.40065527732273, 54.66885092440908, 51.81371401825415, 51.34565878773695, 49.77767376550433, 48.95857711209923, 60.776971682658555, 58.29627896091739, 56.16662766206412, 54.17739293236602, 53.49871284811608, 52.188158202667914, 52.00093611046103, 50.45635384975427, 48.56073016615961, 47.250175520711444, 61.151415867072316, 60.02808331383103, 58.06225134565879, 56.30704423121929, 55.09010063187456, 54.20079569389188, 53.358296278960914, 52.70301895623684, 53.428504563538496, 51.790311256728295, 61.29183243622747]}
+# all={"_without_attack":{"train_loss": [0.1252807784309834, 0.09742209438513659, 0.07278989426693155, 0.08535952889224409, 0.09707955352525141, 0.07154503884881276, 0.09934057234035304, 0.07054111945153942, 0.07282866435186977, 0.07027511819610589, 0.07797958119692334, 0.07450730014005129, 0.06318532331113828, 0.0684236471723281, 0.07826983601088373, 0.0643302706302155, 0.06364701281959831, 0.05664193825045236, 0.06594812614588229, 0.06809216274229402, 0.06446738034718814, 0.053651493564754224, 0.051375579039513744, 0.05154672399682063, 0.05013757950816611, 0.052309241185797016, 0.0540506968701341, 0.05350282117871721, 0.04995783213156694, 0.048244368058029216, 0.05586778874264832, 0.053960465553853365, 0.05318553002894142, 0.05720630808995754, 0.052373226702851806, 0.04640004227402915, 0.04625176661464378, 0.0460953905742425, 0.05406244817375085, 0.05461075934953401, 0.04690211861610349, 0.05021694196844747, 0.04744636542285431, 0.04501832581047782, 0.04821552893232746, 0.04454255953004149, 0.042101731506350985, 0.043201997854622996, 0.04605794766819941, 0.03956587982413299, 0.04384280690736653, 0.04457499277751137, 0.0421911870223546, 0.0437418411275202, 0.042015208516325495, 0.04202706765742059, 0.04496735455778251, 0.04229426422720005, 0.04068845667855255, 0.04505856847386548, 0.04470124124829551, 0.040727966162622475, 0.04118070848141625, 0.04310270623977941, 0.037058768998154905, 0.04445792127331559, 0.04198979077744515, 0.045611549299071866, 0.03687969687576888, 0.035795092530015137, 0.03758594533469299, 0.04366033829507701, 0.040549804229417105, 0.040938271179519904, 0.042481890288810956, 0.03679296865520226, 0.03743105587826852, 0.036951739428609236, 0.03350814963194973, 0.03524786478880131, 0.03643312781846111, 0.03152660275309891, 0.043846164813229435, 0.03576702427541929, 0.04270644246126406, 0.03933881880711511, 0.036621864114661576, 0.03545134261049262, 0.038379858089246634, 0.03543607313125209, 0.042119076335594974, 0.041171170802667995, 0.040578116072739846, 0.03865072032548793, 0.03605181780279291, 0.0351914064417711, 0.03696528720238351, 0.03825749001523422, 0.03471947890255819, 0.04046840616237553], "test_loss": [0.01873986346721649, 0.01795187518596649, 0.01749327528476715, 0.016612403535842896, 0.015582855331897736, 0.01432559642791748, 0.012942187678813935, 0.01256927375793457, 0.011265034091472627, 0.011167864322662353, 0.010905730819702149, 0.00949214208126068, 0.009382576048374176, 0.008574830758571624, 0.007748222124576569, 0.008314536964893341, 0.008173576545715332, 0.008097315746545791, 0.008559942454099655, 0.007890518206357956, 0.00808913598060608, 0.007513561826944351, 0.007504263889789581, 0.007554416644573212, 0.0074902319252490995, 0.0064609851717948915, 0.006354117166996002, 0.006460945850610733, 0.007124226838350296, 0.007118017667531967, 0.006157216829061508, 0.006354311990737915, 0.006029405254125595, 0.007486358183622361, 0.006361633288860321, 0.006555439001321792, 0.006523406594991684, 0.006447110086679459, 0.005719209951162338, 0.006000698155164718, 0.006992172479629516, 0.0063207405745983125, 0.005908599632978439, 0.006139561021327972, 0.006186681753396988, 0.005386266559362411, 0.005850607812404632, 0.0061197901606559755, 0.00538701496720314, 0.005424428910017013, 0.00545900599360466, 0.005431692534685135, 0.006099637591838836, 0.0062327340960502626, 0.005642014652490616, 0.005413393810391426, 0.005554311448335647, 0.005575977087020874, 0.005610974872112274, 0.005934179800748825, 0.005256874144077301, 0.005479650628566742, 0.005182094758749008, 0.005286445075273514, 0.004898989641666412, 0.005544626957178116, 0.005656550562381744, 0.005628841292858124, 0.0051244081437587735, 0.005145332261919975, 0.005115558323264122, 0.005495495557785034, 0.004841230568289757, 0.0050152177035808565, 0.005114789927005768, 0.005529060608148575, 0.005617575579881668, 0.004842384114861488, 0.0051516175985336305, 0.0048248997628688815, 0.004833816683292389, 0.00498144561946392, 0.004959668841958046, 0.005115862512588501, 0.00492893428504467, 0.00498985185623169, 0.004698838165402413, 0.005327022671699524, 0.005103945708274841, 0.0047715742141008375, 0.004583150944113731, 0.00501425401866436, 0.004803160917758942, 0.005671905660629272, 0.0053308781206607815, 0.004617161110043525, 0.004927958458662033, 0.005240088859200478, 0.005024097767472267, 0.0047650525718927386], "test_accuracy": [10.0, 12.67, 19.74, 30.4, 38.9, 36.86, 50.61, 51.05, 58.87, 53.23, 54.01, 61.76, 64.86, 65.93, 70.78, 68.64, 69.7, 69.49, 63.15, 64.43, 63.96, 69.55, 68.22, 69.17, 66.71, 75.92, 75.51, 75.8, 72.45, 72.09, 73.81, 74.61, 77.22, 67.26, 74.69, 72.06, 73.75, 74.42, 76.0, 76.9, 68.15, 73.12, 74.85, 75.17, 73.78, 78.2, 76.42, 74.36, 79.11, 79.0, 78.36, 77.91, 73.69, 75.35, 77.32, 78.06, 78.7, 78.12, 77.59, 74.95, 79.66, 77.59, 78.73, 79.79, 80.32, 76.25, 77.72, 76.39, 79.14, 79.51, 80.14, 79.01, 80.29, 79.23, 78.95, 76.36, 76.49, 80.58, 78.83, 80.57, 79.5, 80.38, 79.45, 78.83, 80.36, 79.33, 80.87, 78.75, 78.56, 80.38, 80.77, 80.05, 81.24, 74.79, 77.28, 80.96, 78.54, 77.39, 78.46, 80.91], "train_accuracy": [20.513283932767035, 19.876009768927297, 22.812399095899256, 40.69789859263543, 45.11506730351715, 41.77855274629468, 56.50652985074627, 50.50161232533142, 63.66064092029581, 59.950903033491144, 62.43362831858407, 65.74340793263904, 72.23520249221184, 71.69601069757076, 74.27926798696416, 70.71970473651835, 71.25662994696043, 75.10340752412843, 70.39718659495242, 68.3348095659876, 66.06174384647476, 74.78391582111988, 73.09919671212404, 73.60974651208488, 72.6086956521739, 78.94324853228963, 78.66426714657068, 76.30285152409046, 73.85165662650603, 74.34222464162583, 78.48970251716247, 77.64491284961491, 79.13879598662207, 69.2824171212757, 79.35867895545314, 79.84585741811175, 77.32368896925858, 79.48520937379946, 80.01361161524501, 81.39641708773541, 75.60975609756098, 76.21136173767752, 80.81534772182255, 77.36964701666368, 80.33303330333034, 82.320959880015, 78.63789348643833, 78.84270870655679, 81.33107727016356, 83.69731800766283, 81.22233930453109, 83.70004490345757, 81.23007687980498, 79.65981573352232, 81.33406034169393, 81.64428164428165, 84.47826086956522, 83.47154165096117, 81.4207650273224, 80.2919708029197, 83.87470997679814, 82.62509928514694, 84.29530201342281, 83.27637761640325, 86.72649093288464, 84.39949431099873, 80.1295484854258, 84.22980086976425, 82.98294327647758, 84.3804537521815, 84.65919202705108, 82.89017341040463, 85.03893214682981, 84.58387235877534, 84.95556037285931, 82.72017837235228, 82.39857766284145, 85.26906697638405, 84.37334276118084, 86.93419551160137, 85.96877502001601, 87.12836656173488, 83.14150304671632, 85.61854736460786, 85.67264573991031, 84.26644871270943, 86.40312771503041, 84.5282350858559, 86.34406397375436, 84.91274160255207, 87.70025839793281, 84.65804066543438, 85.52005438477227, 83.56505499066196, 85.00763941940413, 87.36992450520302, 87.79202931745304, 84.17251341312422, 84.46137491141035, 84.9608647036899], "backdoor_test_loss": [], "backdoor_test_accuracy": []}
+# , "_with_attack": {"train_loss": [0.11388084455849698, 0.09217523318831058, 0.09887498498941584, 0.08555799016637838, 0.10149915561356734, 0.09016060230000777, 0.07688274177357861, 0.07802515233201109, 0.07548322398041052, 0.07072405641778916, 0.08142242926770953, 0.06782249238134623, 0.06730680163424305, 0.06225099202071142, 0.06434370258654531, 0.06770575127159709, 0.06572360714760733, 0.057063528146996424, 0.05954783233593717, 0.053564671492653114, 0.06295454016039762, 0.059324682961638765, 0.05800942676114518, 0.056600433676847264, 0.04926841231396395, 0.05527821710965679, 0.05778058928144425, 0.053210336383451504, 0.05577777197332916, 0.04882410943207627, 0.05601233317958758, 0.05091541632974065, 0.046690520206236, 0.045722449869835806, 0.04867035905106074, 0.05272241020352882, 0.05124258372881, 0.0454965075186607, 0.05248036504539313, 0.04039184475299383, 0.04683919034625775, 0.049725712137373056, 0.048243982953633004, 0.05007790059663427, 0.04427255434158761, 0.048713069004756075, 0.042985524296471594, 0.04669463030571215, 0.04986771935610668, 0.044757998151839354, 0.04590155801670442, 0.045282676925437235, 0.04800435831609522, 0.040468678070610345, 0.04385298993628399, 0.04516930352422294, 0.04020253872723741, 0.044941864592189074, 0.046178481938407304, 0.041551194642215655, 0.038848803246577746, 0.045886555004433165, 0.03598444548261541, 0.041026206613348645, 0.04289398108859558, 0.04539618941910775, 0.04264056403204143, 0.04151003220267043, 0.03860136238324907, 0.04441375399014628, 0.04034387629247673, 0.042321724699349195, 0.03462074944635038, 0.03737434336030518, 0.0386014263502821, 0.03807151763936798, 0.039579627673334464, 0.037220472200262325, 0.037331486723522836, 0.033480604909999725, 0.04073981607298325, 0.03358746571200955, 0.03621341091707228, 0.03996634988019521, 0.037406033750060796, 0.03663862570109123, 0.037381384824902, 0.03400927144645368, 0.03354176064239221, 0.03237961983675206, 0.03349839531467372, 0.03425543016350613, 0.03922430338977295, 0.042237131901477765, 0.03543396415492414, 0.039762330630486634, 0.034889670535384484, 0.033189174376878613, 0.03638660186645892, 0.0311256394524091], "test_loss": [0.017946914649009705, 0.01756287875175476, 0.016939151549339296, 0.017300402450561525, 0.01576265995502472, 0.015408744585514068, 0.014976538097858429, 0.01259388520717621, 0.012281618022918701, 0.01566720085144043, 0.011598931324481964, 0.009840813326835632, 0.009921973359584809, 0.009132604897022247, 0.009821049189567567, 0.009116015672683716, 0.008397632133960723, 0.008872477030754089, 0.008793489587306977, 0.010039829325675964, 0.0073506894111633305, 0.008094397640228271, 0.006909790062904358, 0.007425745230913162, 0.006836298471689224, 0.006798752868175506, 0.008013131308555603, 0.007361459529399872, 0.006971509724855423, 0.009529452955722809, 0.007292890810966492, 0.006124896478652954, 0.006528359949588776, 0.006385119897127151, 0.006223866057395935, 0.005926008158922196, 0.005801262110471725, 0.006520108330249786, 0.00616519730091095, 0.007737939184904098, 0.00640998061299324, 0.005761722505092621, 0.005434030026197433, 0.006662865906953812, 0.006820247918367386, 0.006073400312662125, 0.006205206644535065, 0.006084989339113236, 0.006201096212863922, 0.008270239078998565, 0.005760375434160233, 0.005910332524776458, 0.005706694412231445, 0.005869086092710495, 0.0057826163291931154, 0.005743516862392425, 0.005307218605279922, 0.005507039296627045, 0.005586506527662277, 0.007017270761728287, 0.0055348081469535826, 0.005620185220241547, 0.005323079857230187, 0.005552519631385804, 0.005377266377210617, 0.005576376748085022, 0.005561141061782837, 0.005116114941239357, 0.0051049626559019085, 0.006578046172857285, 0.005089732149243355, 0.0050703039169311525, 0.00526347661614418, 0.005108841788768768, 0.0051017523169517515, 0.005114426875114441, 0.005069406715035438, 0.005205585178732872, 0.0051248321026563645, 0.006633793205022812, 0.0055003897488117215, 0.005219807478785515, 0.005175902926921844, 0.0047029045104980466, 0.004933711004257203, 0.0045661817759275435, 0.004776694884896278, 0.0048604116171598434, 0.004850495001673698, 0.006478584098815918, 0.004842981797456741, 0.004811103260517121, 0.005011964893341064, 0.004711525559425354, 0.004615911445021629, 0.0046953065097332, 0.004923999631404876, 0.004995693176984787, 0.0048749953180551526, 0.005687134140729904], "test_accuracy": [19.03, 15.38, 25.92, 21.09, 34.62, 43.2, 42.19, 57.03, 57.18, 19.72, 59.22, 61.59, 61.82, 62.35, 59.2, 68.83, 69.19, 63.26, 65.37, 63.16, 73.01, 65.59, 72.24, 69.84, 72.83, 73.58, 65.81, 70.18, 72.23, 65.8, 70.0, 75.0, 74.61, 72.32, 75.47, 75.2, 77.08, 71.28, 73.54, 70.94, 73.87, 76.55, 77.51, 72.1, 71.08, 77.2, 73.62, 73.94, 73.49, 71.31, 77.26, 76.13, 75.94, 76.29, 76.68, 77.24, 78.63, 77.35, 76.72, 75.01, 78.06, 75.89, 78.24, 77.66, 77.75, 78.21, 77.08, 79.92, 78.82, 77.57, 79.74, 78.93, 78.25, 78.9, 78.39, 77.79, 79.73, 77.39, 79.06, 76.21, 77.91, 78.71, 77.82, 80.2, 79.64, 80.86, 80.02, 79.95, 79.79, 74.97, 79.74, 80.24, 78.35, 80.04, 80.4, 80.24, 79.88, 79.9, 80.18, 77.87], "train_accuracy": [20.11292834890966, 22.009821115398108, 29.130156016242786, 25.265622186205654, 36.43146450650181, 47.27312390924956, 49.13210305134296, 59.826118456801304, 60.354196015294825, 24.862365117815457, 60.65259117082534, 67.00847958982449, 65.65832848555361, 67.47800037446171, 64.1382130443672, 72.15612606971405, 70.8185053380783, 67.10298363811357, 69.65411809669406, 65.23908523908524, 74.93398334348974, 69.24872665534805, 73.64891518737673, 70.276008492569, 76.58663688814443, 78.34782608695652, 70.8491761723701, 76.63463350292066, 73.40894770006301, 64.6096654275093, 78.07256235827664, 79.1334559574903, 77.65554553651938, 78.79118878572729, 78.20949329060684, 79.05201847760594, 80.8283103247132, 75.60235249478278, 77.9934613812832, 73.7198645148436, 80.2608526377263, 78.3600493218249, 80.07283026502124, 78.98745519713262, 75.59303187546331, 81.11671190383869, 79.29532858273951, 76.22540250447227, 81.00545950864422, 74.44468880580602, 81.51946421096693, 81.10804500106134, 81.17550349362926, 81.56505914467698, 80.5789164873851, 82.45998684499014, 81.84877276326208, 81.6941074523397, 80.50847457627118, 77.61194029850746, 82.46321306188268, 79.59445037353255, 83.4480122324159, 81.22827346465817, 83.43962585034014, 82.29951262738147, 81.83457290962362, 81.67425291905799, 84.23947535434736, 76.74714104193139, 84.58942632170978, 84.45317345462064, 83.20112815089018, 83.29281000347343, 84.01579713157348, 85.14251061249242, 83.4965034965035, 85.55164562202442, 85.31142219517143, 79.66855295068714, 82.39983217956787, 86.38551896784132, 84.65168963564281, 85.92200925313946, 86.5979381443299, 86.84489709314661, 87.62541806020067, 85.08949679162446, 86.48648648648648, 80.38141610874416, 86.4136846244593, 84.64133522727273, 86.21948437143509, 86.62735849056604, 86.4936657131181, 86.00133958472874, 86.98733089972085, 86.42344497607655, 85.43650793650794, 83.68924889543446], "backdoor_test_loss": [0.02035635542869568, 0.017672666931152343, 0.021698704266548158, 0.019938273453712463, 0.02146624324321747, 0.021643255472183227, 0.020015637969970704, 0.023154831147193907, 0.022595325899124146, 0.008550325119495392, 0.02403405442237854, 0.022297055363655092, 0.027668843483924867, 0.022489186811447143, 0.02164952657222748, 0.02156012091636658, 0.02531903223991394, 0.029946950387954713, 0.030563764977455138, 0.012635911893844604, 0.02637246036529541, 0.02426691153049469, 0.02925299768447876, 0.024430638933181762, 0.03052748372554779, 0.024136224937438965, 0.03351198773384094, 0.026778493857383727, 0.03260307331085205, 0.015456887078285218, 0.021558890867233275, 0.0237959623336792, 0.02846464126110077, 0.024514722394943237, 0.031160579061508178, 0.027311859655380247, 0.031528289008140566, 0.032423638796806334, 0.03119774408340454, 0.013454596519470214, 0.026229639649391175, 0.027330278253555298, 0.03350509910583496, 0.034449054622650147, 0.030824341297149657, 0.03168677308559418, 0.03461025788784027, 0.029796031165122985, 0.038669759511947635, 0.013502086973190308, 0.02374106044769287, 0.026870901846885682, 0.029999179673194886, 0.023850857853889466, 0.03407115683555603, 0.030806685662269594, 0.03558099544048309, 0.03399374804496765, 0.028195462059974672, 0.01540606722831726, 0.023849525809288025, 0.03042130765914917, 0.028379789710044862, 0.03244146029949188, 0.027052747750282286, 0.031116471600532533, 0.03283098473548889, 0.03536115689277649, 0.03126340095996857, 0.015376589870452881, 0.024994873929023744, 0.026173068833351135, 0.03091112720966339, 0.031223453783988952, 0.03632264370918274, 0.03612359328269959, 0.03183250534534454, 0.03505679783821106, 0.03382583014965057, 0.014496056973934173, 0.025086258125305176, 0.0272166211605072, 0.02710488679409027, 0.0346208193063736, 0.026363193249702453, 0.030324762749671937, 0.03390371277332306, 0.03574615299701691, 0.03134391894340515, 0.013896755194664002, 0.022519228839874266, 0.029933043766021728, 0.031212251949310304, 0.032715825510025026, 0.03144959990978241, 0.03234345400333404, 0.03340684041976929, 0.03517869625091553, 0.03267859539985657, 0.014543625724315643], "backdoor_test_accuracy": [0.0, 4.77, 0.02, 2.47, 0.0, 3.0, 17.79, 3.79, 6.69, 91.56, 3.47, 12.92, 0.77, 16.91, 18.23, 11.11, 10.01, 6.01, 11.56, 39.92, 9.56, 14.62, 10.96, 14.58, 8.39, 18.49, 6.23, 15.4, 8.69, 30.09, 25.54, 19.0, 10.6, 20.2, 7.26, 14.86, 10.29, 10.51, 10.89, 38.88, 16.02, 12.95, 10.42, 5.71, 13.49, 7.97, 9.13, 14.0, 7.07, 39.57, 17.82, 13.69, 11.43, 21.68, 7.07, 12.66, 9.02, 9.02, 17.16, 33.55, 17.77, 13.06, 15.56, 12.56, 19.01, 11.85, 11.36, 11.18, 17.01, 36.63, 18.46, 20.01, 11.22, 15.19, 8.74, 11.77, 14.09, 11.13, 12.16, 41.8, 18.18, 18.32, 20.78, 11.39, 22.83, 17.16, 12.69, 12.29, 19.26, 42.02, 26.31, 15.95, 15.44, 15.25, 17.59, 16.35, 15.04, 11.92, 15.55, 42.53]}
+# }
+
+all = {"_without_attack": results}
+plt.rcParams.update({'font.size': 8})
+ax = plt.subplot()
+
+rounds = len(all["_without_attack"]["test_accuracy"])
+
+print(rounds)
+
+x_axis = np.arange(1, rounds + 1)
+i = 0
+colours={"test_accuracy_without_attack": "g",
+         "test_accuracy_with_attack": "b",
+         "backdoor_accuracy": "r"}
+
+for results in all.keys():
+    for metric in all[results].keys():
+        values = all[results][metric]
+        if len(values) > 0 and "accuracy" in metric and "train" not in metric:
+            label=metric+results
+            if "backdoor" in metric:
+                label = "backdoor_accuracy"
+            y_axis = np.array(values)
+            ax.plot(x_axis, y_axis,colours[label], label=label)
+            
+            i += 1
+    ax.legend(loc='lower right')
+    ax.set(xlabel='Number of Rounds', ylabel='Accuracy')
+ax.grid()
+ax.set_yticks([0,10,20,30,40,50,60,70,80,90, 100])
+# ax.set_xticks([0,10,20,30,40,50,60,70,80,90,100, 110])
+
+plt.title("Accuracy with no attack for the Blood Cell dataset")
+plt.show()
+
